backend/requirements.txt
fastapi
python-dotenv
pyowm
ctransformers
progressbar
langchain
chainlit